Virtual memory manager.
The first 512GB of in virtual memory as an identity map, this allows
for easy access when we're doing some guest <-> host translation in
the future.

After the 512GB of virtual memory, virtual memory can then be allocated
(using the vmem_alloc call), again this is the physical memory from the
first 512GB however this will allow for storage of large contiguous data
without having to worry about physical memory fragmentation.