Generic VMCS setup for the VMM.
EPT configuration for the VM is configured, retrieving the link pointer
from the EPT module and setting up a singular VPID that will be re-used
for each vCPU.

MSR trapping bitmap is enabled, currently this bitmap is initialised
to zeros, as such EVERY MSR read/write will be VMExit'd, however in
the future we have the ability to amend this so it doesn't happen.

Processor based controls are also configured for the environment,
RDTSCP, XSAVES/XRESTORES and unrestricted guest are configured
among other things such as EPT and VPID.

VMM is configured so that entry/exit will ALWAYS land us in long/x64
mode.