Guest VMCS setup for the VMM.
Configuration of the segment selectors, GS, GDTR, IDTR etc.

For the segment selectors we re-use the host GDTR and it's selectors
this is for each of use as unfortunately to be able to VMENTER successfully
the TR, TSS & LDTR need to be present which UEFI (the original GDTR) does
not provide for us.

However, when the OS is loader later on this will get replaced anyway so
this *SHOULD* be okay...

CR3 again has to be re-used from the host environment, as technically our shim
or rest of the driver may try access our dynamically allocated memory in the guest
during exit of the hyperjacking routine. Again this *SHOULD* be okay for now as when
a guest OS is loaded CR3 will be written and the guest will be non the wiser that
this was technically from cascade.

Guest RSP & RIP are set so that the host_to_guest shim will be called, this will
then in the future jump straight back to our DXE driver code and continue the
"driver load" in UEFI which finished the hyperjack.