Unfuck dynamic self-relocating in gnu-efi.
So it turns out all of the issues with relocations not working correctly was due
to not setting -shared in the linker flag. This originally was done so that I could
actually get linker errors when a function wasn't included etc.

HOWEVER this turned out to be a really fucking stupid idea, as _DYNAMIC is used for
the self-relocating code within gnu-efi so that the dynamic relocations can take
place. I REALLY wish it wasn't done this way and instead matched EDK or VisualUEFI.

Now though I don't have to worry about manually creating actual addresses from their
offsets, therefore RIP for guest & host have been used to use the absolute addr.
IDT has also been modified so that the vector table is stored within .data and contains
absolute addresses of the ISR stubs (FINALLY)!

This has been a nightmare, really hope I don't get issues with functions being missing
and causing null pointer execution errors in the future however that may be easier to
find.