Basic VMEXIT handling.
Successfully retrieves the vcpu_context that causes the VMEXIT and stores
within the vcpu context. The exit reason is then retrieved and a skeleton
for the actual handling of this will be implemented next.

Fixed bug in __vmread intrinsic where the dest and source were mixed up,
now the return value of a __vmread is stored in the resultant RAX of the
function.