Added debug script for loading the hypervisor and then a windows 10 guest.
To be able to run this successfully, you need to prevent the UEFI
environment from trying to boot from the win10 guest first. So as
soon as the display is loading spam ESC in the console until you
get to the boot menu screen. Adjust to boot from EFI shell then
continue from there.