Fixed issue where we were receiving interrupts from PIT in host mode...
This would then cause an exception in our host interrupt handler.
From now on if an interrupt happens from within host mode we forward
the interrupt on to the guest.

If the interrupt number is below 0x20 this means it's came internal,
we should actually deal with this ourselves rather than forwarding
to guest as it's possible we could leak critical info to guest.

TODO:
	Make this SMP safe.
	Make this work with per vCPU interrupt handling.