Trapping on MSR read/write via MSR bitmap and MTF exits.
Trapping on MSR read/write is not as simple as setting the whole MSR
bitmap to 1's and then attempting to rdmsr/wrmsr within host context.

This is due to the posibility that we may accidentally leak some MSR's
that are being used strictly for the hypervisor, such as GS_BASE as we
use this as a hack to store our vCPU context on VMEXIT to host.

To get around this, we trap on ALL MSR's unless they're explicitly allow-
listed. Once done we check RPL and MSR range, if these are okay we then
disable MSR trapping on the selected vCPU, once this is done we then
enable MTF trapping so that we catch as soon as the MSR is read/written.

Once done we then re-enable the MSR trap via bitmap and disable MTF.
This is VERY wasteful of cpu time but if we want to log MSR read/writes
this is the best way.

NOTE: This MAY not work all the time, imagine if there is a context switch
at the time of the VMENTER after the initial MSR r/w trap. We truly didn't
read it, so we may see multiple events show up for technically the same
read, however it's good enough for our usage.