Started defining the VMCALL and Plugin interfaces for cascade.
The VMCALL interface will allow a guest application (that knows the secret
key) to be abel to query/action the hypervisor, for example to load a plugin
or to gather events from the event log.

The plugin interface is created so that the base of cascade can remain unchanged
however we can create plugins for logging, hooking or any other manner of things.