Plugin image validation.
Plugin images are going to be based of Windows x64 DLL images.
This is due to ease the development process as it's already a defined
format and also means plugins in the future can be created in tree
using the GCC/MinGW/Makefile environment or out of tree using MSVC.

Currently checks the validity of the PE headers of the plugin.