Nested MSR read/write support.
This will be used for trapping on certain MSRs we may want to spoof to
the guest environment when we are running nested virtualization. For
example the VMX_REVISION_ID.

Alongside this, we fix the broken logic for the revision check when
a VMXON takes place, now this checks against what we set rather than
what we were relaying from the host.