Functionality to hide hypervisor (EFI RT Driver) from guest.
This currently is NOT enabled, we will need to write a
VMCALL interface that enables this via either a plugin or
a general usermode application.

Utilize EPT/SLAT to hide the runtime driver from the guests
perspective. Redirecting all pages (visible to the guest) to
a dummy page that is zero'd. In the future we can also set
to no RWX and record if the guest does try to probe that
memory.