Tidy up directory and build structure for usermode.
Create a new "usermode" folder which can hold plugins and
other usermode applications we may create. This is due to
the fact that we may want to create other EXE's that talk
with the hypervisor rather than just a plugin-loader (and
the respective plugins).

This is pre-emptive as we will be creating more "test"
utilities that can be used to verify functionality is not
regressing.