Refactor the plugin interface to allow for registering VMEXIT callbacks correctly.
Due to prior commits where we adjusted the VMEXIT handler
the exit events would not be called, as such now we just
create a shim to the standard VMEXIT handler which can be
called by plugins.

Note: This MAY/MAY NOT work, there are no plugins currently
using the VMEXIT events, maybe we should write one soon.