Refactor the "interface" folder layout.
This is part of an effort to neaten up cascade as a whole.

Creating a "hypervisor" folder for interface to the HV/VMM
specifics, for example an EFI runtime driver which inherits
the hypervisor library.

Creating a "usermode" folder which gives interface for the
VMCALL logic (which in the future will be refactored for
VMCALL event register/deregister like we do for VMEXIT's
currently.