Make the hypervisor a static library
Instead of messing around trying to make this project plugin
based etc. Let's just make "cascade" a hypervisor library,
therefore people can use it as a submodule and build around
it for whatever their needs are.

An example is given for a UEFI runtime driver, basically
providing the implementation hooks for running on all LPs
and getting processor index via the MP_SERVICES_PROTOCOL.

Also:
elf_x86_64_efi.lds is a modified version of the EFI linker
script which keeps .bss in it's own section. This can be
originally found in commit:
62a77d79bc59db96a9925476f7b6e17c615ed00e