Ensure EFER.NX is enabled during vmem initialisation.
Some UEFI implementations (looking at you VMWARE) do not set
the NX bit by default in UEFI / Long mode. Therefore as we
utilise NX capabilities within our vmem allocator we need to
ensure it's enabled.

If we do not enable it we get a cryptic page fault when we
try to read/write to a page that is virtually allocated with
it's PTE set to NX.