Add in new wait_for_debugger utility to aid GDB debugging
There are scenarios when we want to debug issues with GDB,
for example we want to catch a specific part of code running
and see what happens (e.g a corruption).

This wait_for_debugger utility helps with this, it basically
causes a vCPU to go into a spin loop until we manually set
EAX to zero, allowing us to catch execution.

Also lets add in a commands file for GDB which we can use
for manually loading some commands when we start the debug,
(e.g setting syntax to Intel since AT&T is trash).