Remove the shim_host_to_guest assembly
This is redundant code when we can jump straight to a C
function. So let's remove it.

While we're here let's also adjust how this shim works.
Since VMX-root is always in long mode we can (ab)use the
GS pointer to hold a pointer to the vCPUs specific context
structure. This simplifies the code A LOT.