Fix broken MTRR code on newer processors
MTRR has been broken since the beginning, we were not taking
into account the fixed MTRR ranges but also we assumed that
by default any out of range areas would be WB.

This is NOT true, this has caused me so much pain over the
past few weeks trying to get this project to boot on a 13th
gen Intel.

Now we properly deal with the MTRRs.