Fix running_as_guest bug & guest_context in hyperjack bug
The running_as_guest flag is incorrectly named, this should
be to indicate the hypervisor has been launched. If we clear
it in the exit handler this can cause a race if a VMEXIT
occurs before the hypervisor has finished it's hyperjack
restore.

Also the guest_context is bug prone if we use it in the VMM
main module for the hyperjack, hence let's create another
context and use it ONLY for that.