Force SS segment to use full 4GiB limit.
This bug nearly made me quit this project completely. On a
ASUS Z790 motherboard the UEFI environment will put the AP
startup shim in a high place in memory. When the UEFI
ExitBootServices call is made some code is ran on all APs
which puts the AP into protected mode rather than long mode.

As soon as we go and perform an operation that involves the
stack we were getting #SS(0) which then results in a triple
fault since the AP has no IDTR that is deliverable to.

We didn't catch this on older motherboards as the MpInit
module did not put the shim code in a higher part of mem.
Ofcourse this didn't happen in QEMU too as the UEFI put the
shim in a sane part of memory.

Stupidest fucking bug in the world, maybe I should try fix
this properly but I've spent TOO long on this and I can't
be bothered now.

This project now successfully boots on a 13th gen Intel and
Windows/Ubuntu/whatever now succesfully are hyperjacked.